+++ fixtures/recall/legacy-cypress-steps/common.cjs
@@ -1,0 +1,3 @@
+const legacySteps = require("cypress-cucumber-preprocessor/steps");
+
+legacySteps.Given("legacy Cypress registration", () => openWorkspace());